+++ Parser.cpp
@@ -21,7 +21,7 @@ using namespace std;
 void Parser::parse(){
     int index = 0;
     auto data = Data::getInstance();
-    while (index < this->lexer.size()) {
+    while (index < (int)this->lexer.size()) {
         Command* command1 = (Command*)data->getCommandMap(this->lexer[index]);
         if (command1 != nullptr){
             index += command1->execute(index, this->lexer);
